PyNMRSTAR automatically retries fetching an entry from the BMR API using an exponential backoff if rate limited.
PyNMRSTAR now lists the package requests as a requirement, which allows it to significantly speed up fetching entries from the database. It will still work if requests is not installed though, as in the case where you have checked out the code locally and don't have requests installed - you just won't get the enhanced performance.

Closes #91.